chore: move test components' enums to common
